rename util classes
