+++ src/main/java/com/researchspace/core/testutil/JakartaValidatorTest.java
@@ -1,29 +1,28 @@
-package com.researchspace.core.testutilJU5;
+package com.researchspace.core.testutil;
 
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
 import java.util.Set;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
+import jakarta.validation.ValidatorFactory;
 
-import org.junit.jupiter.api.BeforeAll;
-
-import lombok.extern.slf4j.Slf4j;
+import org.junit.BeforeClass;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 /**
- * Utility methods for testing Javax Validator annotations on a class.
- * PRovides a {@link Validator} instance to subclasses.
+ * Utility methods for testing Jakarta validation annotations on a class.
+ * Provides a {@link Validator} instance to subclasses.
  *
  */
-@Slf4j
-public abstract class JavaxValidatorTestJU5 {
+public abstract class JakartaValidatorTest {
 	
 	protected static Validator validator;
+	Logger log = LoggerFactory.getLogger(JakartaValidatorTest.class);
 	
-	@BeforeAll
+	@BeforeClass
 	public static void beforeClass() {
 		ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
 		validator = factory.getValidator();
